place unit tests of config in standard location for zainod
